+++ .eslintrc.json
@@ -37,7 +37,7 @@
     "import/no-extraneous-dependencies": [
       "error",
       {
-        "devDependencies": [".storybook/**", "stories/**"]
+        "devDependencies": [".storybook/**", "stories/**", "**/*.stories.tsx"]
       }
     ]
   }
