+++ tsconfig.json
@@ -26,7 +26,8 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    ".storybook/**/*.ts",
+    ".storybook/*.ts",
+    ".storybook/*.tsx",
     "cypress.config.js",
     "cypress/**/*.js",
     "cypress/**/*.ts",
